requirements.txt
flask
flask-sqlalchemy
mysqlclient
python-dotenv